Moves manager/owner check into checks.py and uses it consistently
