+++ src/cogs/registrator.py
@@ -19,18 +19,9 @@ load_dotenv()
 EVENT_NAME = os.getenv('EVENT_NAME')
 EVENT_GUILD_ID = int(os.getenv('EVENT_GUILD_ID'))
 EVENT_CONTACT_EMAIL = os.getenv('EVENT_CONTACT_EMAIL')
-EVENT_BOT_CHANNEL_ID = int(os.getenv('EVENT_BOT_CHANNEL_ID'))
-BOT_MANAGER_ROLE_ID = int(os.getenv('BOT_MANAGER_ROLE_ID'))
-GUILD_OWNER_ID = int(os.getenv('GUILD_OWNER_ID'))
 
 IS_PROD = os.getenv('IS_PROD')
 LOGGING_STR = os.getenv('LOGGING_STR')
-
-def is_owner_or_botmgr():
-    async def predicate(ctx):
-        role = discord.utils.get(ctx.guild.roles, id=BOT_MANAGER_ROLE_ID)
-        return (role in ctx.author.roles) or ctx.author.id == GUILD_OWNER_ID
-    return commands.check(predicate)
 
 async def sync_server_roles(guild: discord.Guild, member: discord.Member, participant: models.Participant):
     # Add Discord Roles.
@@ -350,7 +341,7 @@ class Registrator(commands.Cog):
                 )
 
     @commands.guild_only()
-    @checks.is_in_channel(EVENT_BOT_CHANNEL_ID)
+    @checks.is_in_bot_channel()
     @registrator_group.command(name="reset", description="🚫 [RESTRICTED] Reset a users\'s roles.")
     async def _reset_user(self, ctx,
                           member: Option(discord.Member,
@@ -402,7 +393,7 @@ class Registrator(commands.Cog):
             self.log.info(f"**`ERROR:`** _reset_user_error[{ctx.author.name}]: {error}")
 
     @commands.guild_only()
-    @checks.is_in_channel(EVENT_BOT_CHANNEL_ID)
+    @checks.is_in_bot_channel()
     @registrator_group.command(name="connect_account", description="🚫 [RESTRICTED] Register a user.")
     async def _add_participant(self, ctx,
                                member: Option(discord.Member,
